requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
scipy>=1.12.0
openai>=1.14.0
pydantic>=2.6.0
python-multipart>=0.0.9
requests>=2.31.0
pyyaml>=6.0.1
